refactor: ApplyPage
