+++ src/assets/images/apply/requiredInput.png
