+++ src/pages/ApplyPage/components/Form/components/styles/BasicInfo.css
