+++ src/pages/ApplyPage/components/Form/components/styles/Required.css
